restapi/restapi/requirements.txt
Flask==2.2.2
Flask-PyMongo==2.3.0
PyJWT==2.4.0
pymongo==4.1.1
gunicorn==20.1.0
python-dotenv==0.19.2